requirements.txt
arxiv==1.4.8
nltk==3.8.1
sentence-transformers==2.2.2
torch==2.1.0
scikit-learn==1.3.0
numpy==1.24.3
requests>=2.31.0
pyyaml>=6.0 